+++ src/main/java/com/cydeo/accountingsimplified/service/DashboardService.java
@@ -1,14 +1,12 @@
 package com.cydeo.accountingsimplified.service;
 
-import com.cydeo.accountingsimplified.dto.CompanyDto;
-import com.cydeo.accountingsimplified.dto.InvoiceDto;
-
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 
 public interface DashboardService {
 
-    Map<String, Integer> getSummaryNumbers() throws Exception;
+    Map<String, BigDecimal> getSummaryNumbers() throws Exception;
     List<String> getExchangeRates();
 
 }
