+++ src/main/java/com/cydeo/accountingsimplified/service/ReportingService.java
@@ -1,12 +1,14 @@
 package com.cydeo.accountingsimplified.service;
 
 import com.cydeo.accountingsimplified.dto.InvoiceProductDto;
+
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 
 public interface ReportingService {
 
     List<InvoiceProductDto> getStockData() throws Exception;
-    Map<String, Integer> getMonthlyProfitLossDataMap();
+    Map<String, BigDecimal> getMonthlyProfitLossDataMap();
     
 }
